trying to fix strange compiler bug

+++ samples/cpp/tutorial_code/calib3d/real_time_pose_estimation/src/RobustMatcher.h
@@ -23,7 +23,7 @@ public:
     extractor_ = cv::ORB::create();
 
     // BruteFroce matcher with Norm Hamming is the default matcher
-    matcher_ = cv::makePtr<cv::BFMatcher>(cv::NORM_HAMMING, false);
+    matcher_ = cv::makePtr<cv::BFMatcher>((int)cv::NORM_HAMMING, false);
 
   }
   virtual ~RobustMatcher();
